Update hierarchy.R
Line 38 which adds an additional class type to the data structure does not play well with other functions (such as gghighlight), so removal of the line keeps the data structure as a data.frame which avoids downstream errors.  I only commented the line out, but it could likely be removed. Unless the removal of this line results in conflicts in other scrabble specific functions. Removal fixed my problems.